+++ packages/datadog_gql_link/test/datadog_gql_link_test.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 
 import 'package:datadog_common_test/datadog_common_test.dart';
+import 'package:datadog_common_test/uri_matchers.dart';
 import 'package:datadog_flutter_plugin/datadog_flutter_plugin.dart';
 import 'package:datadog_flutter_plugin/datadog_internal.dart';
 import 'package:datadog_gql_link/datadog_gql_link.dart';