+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+
+
 ## 2.3.0
 
 * Add `batchProcessingLevel` configuration option.