+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `error.source_type` to logs when a stack trace is provided.
 * Fix default console print functions. See [#575] and [#574]
+* Loosen the restriction on the `js` package to allow `<0.8`. See [#572]
 * Add support for global attributes for logs.
 
 ## 2.3.0
@@ -255,5 +256,6 @@ Release 2.0 introduces breaking changes. Follow the [Migration Guide](MIGRATING.
 [#543]: https://github.com/DataDog/dd-sdk-flutter/pull/543
 [#554]: https://github.com/DataDog/dd-sdk-flutter/pull/554
 [#558]: https://github.com/DataDog/dd-sdk-flutter/issues/558
+[#572]: https://github.com/DataDog/dd-sdk-flutter/issues/572
 [#574]: https://github.com/DataDog/dd-sdk-flutter/issues/574
 [#575]: https://github.com/DataDog/dd-sdk-flutter/issues/575